Move firestore integration tests to separate folder
