+++ .gitignore
@@ -49,5 +49,4 @@ yarn-error.log*
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-firebase-debug.log
-firestore-debug.log
+*.log
