+++ firebase-it/firestore.rules
@@ -4,10 +4,6 @@ service cloud.firestore {
   match /databases/{database}/documents {
     match /users/{userId}/{document=**} {
       allow read, write: if request.auth.uid == userId;
-
-      match /users/{userId}/budgets/{budgetId} {
-        allow write: if request.resource.data.identifier == null;
-      }
     }
   }
 }
