+++ firebase-it/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "cloud-firestore-emulator-javascript-quickstart",
+  "version": "1.0.2",
+  "description": "Cloud Firestore emulator testing",
+  "scripts": {
+    "format": "prettier --write **/*.js",
+    "test": "mocha --timeout=10000"
+  },
+  "devDependencies": {
+    "@firebase/testing": "0.10.0",
+    "filesystem": "1.0.1",
+    "mocha": "5.2.0",
+    "prettier": "1.15.2"
+  }
+}
